improved gallery code
