+++ app/Livewire/Images/View.php
@@ -6,23 +6,10 @@ use Livewire\Component;
 
 class View extends Component
 {
-    public $sort;
 
     public function render()
     {
         return view('livewire.images.view');
     }
 
-    public function mount(){
-        $this->sort = "by newest";
-    }
-
-    public function sortChanged(){
-        $this->dispatch('updateSort', $this->sort);
-    }
-
-    public function updateGallery(){
-        
-        $this->dispatch('updateGallery');
-    }
 }
